server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
attrs==25.3.0
branca==0.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.116.1
filelock==3.19.1
folium==0.20.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
geographiclib==2.1
Geometry==0.0.23
geopandas==1.1.1
geopy==2.4.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pip==23.0.1
polars==1.33.1
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pygeoif==1.5.1
pyogrio==0.11.0
pyparsing==3.2.4
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.0
setuptools==66.1.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
triton==3.4.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.199
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
xyzservices==2025.4.0
yarl==1.20.1
